+++ src/web/topic/components/Score/Score.tsx
@@ -5,7 +5,8 @@ import { useSessionUser } from "../../../common/hooks";
 import { usePerspectives } from "../../../view/store/store";
 import { useFlowZoom } from "../../hooks/flowHooks";
 import { useUserScores } from "../../store/scoreHooks";
-import { playgroundUsername, useOnPlayground } from "../../store/store";
+import { playgroundUsername } from "../../store/store";
+import { useOnPlayground } from "../../store/topicHooks";
 import { BackdropPopper, CircleDiv, ScorePopper } from "./Score.styles";
 import { ScoreButton, buttonDiameterRem } from "./ScoreButton";
 import { ScoreCompare } from "./ScoreCompare";
