+++ src/web/topic/store/topicHooks.ts
@@ -1,0 +1,6 @@
+import { useTopicStore } from "./store";
+import { isPlaygroundTopic } from "./utils";
+
+export const useOnPlayground = () => {
+  return useTopicStore((state) => isPlaygroundTopic(state.topic));
+};
