+++ src/web/topic/store/userHooks.ts
@@ -1,8 +1,9 @@
 import { useTopicStore } from "./store";
+import { isPlaygroundTopic } from "./utils";
 
 export const useUserCanEditTopicData = (username?: string) => {
   return useTopicStore((state) => {
-    if (!state.topic) return true;
+    if (isPlaygroundTopic(state.topic)) return true;
     if (!username) return false;
 
     return state.topic.creatorName === username;
